Added test sequence for measuring braking distance
Note: in this test sequence, we're measuring the distance it takes for the robot to stop given the rate of the drive encoders. Assuming the robot is moving straight, the encoder rate is directly related to how fast the bot is moving (speed = v). Given the speed v, we expect its kinetic energy to be proportional to v^2, which then is proportional to the amount of work (F * d) we need to do to stop the bot. Therefore we should expect a parabolic, not linear, function of stopping distance versus encoder rate. (d versus v^2 should be linear)