Added Rotate Brake test sequence and Prefs class
The Prefs class is basically what you can use instead of Initialization.prefs, and when you use Prefs.getX(...) [getDouble, getBoolean, ...] it checks if the key exists, and if it doesn't, it creates it. That way we're making sure we can see all the preferences we're accessing in the code, because if one goes missing from the dashboard, then we're actually using the hardcoded backup value.